Phase 2 Complete: Full Strava Integration with Real Data Validation
✅ Implementation Complete:
- M5 (Strava Integration): OAuth 2.0 flow, activity sync with pagination, rate limiting
- M7 (Notes Analyzer): Deployed in real pipeline, fixed for production use
- M6 (Normalization): Validated with 17 real activities
- M8 (Load Engine): Two-channel model confirmed with real multi-sport data

✅ Real Data Validation (17 Activities):
- 7 climbing sessions: 315 AU systemic, 52 AU lower-body (upper-body dominant)
- 5 running sessions: 301 AU systemic, 301 AU lower-body (balanced load)
- 5 yoga sessions: 20 AU systemic, 6 AU lower-body (recovery-focused)
- Discovered: Strava HR as float (158.1), not int; workout_type values 28, 31 exist
- RPE extraction working: User input > HR-based > text > duration heuristic

✅ Schema Corrections:
- HR fields changed from Optional[int] to Optional[float] (Strava returns 158.1, not 158)
- M7 field references fixed: has_polyline, duration_seconds // 60
- None profile handling added in M7 for graceful degradation
- Strava workout_type enum updated with observed values

✅ Documentation Updated:
- docs/specs/modules/m05_strava.md: Added TEMPO_RUN=28, VO2_MAX_RUN=31; HR float comment
- v0_product_requirements_document.md: HR examples updated to floats
- CLAUDE.md: Added validated real-world load calculations
- IMPLEMENTATION_PROGRESS.md: Phase 2 validation results, progress 67%→72%

✅ Test Coverage:
- M5 (Strava): 28 unit tests (OAuth, fetching, mapping, deduplication, errors)
- Integration: Full M5→M7→M6→M8 pipeline tested with real Strava data
- Files created: oauth_helper.py, get_strava_token.py, sync_strava.py, test_pipeline.py
- STRAVA_TESTING_RESULTS.md: Comprehensive test report with real data insights

✅ Production Readiness:
- All critical paths validated with real data
- Error recovery implemented (partial syncs)
- Deduplication working (primary key + fuzzy match)
- Rate limiting respected (26.5s for 17 activities)
- Atomic writes ensure data consistency

✅ Repository Cleanup:
- Test utilities excluded from git (oauth_helper.py, get_strava_token.py, sync_strava.py, test_pipeline.py, data_samples/)
- .gitignore updated with testing and local data patterns
- DOCS_UPDATES.md removed (changes integrated into documentation)
- Ready for Phase 3 (M9 Metrics Engine: CTL/ATL/TSB/ACWR)

+++ sports_coach_engine/schemas/activity.py
@@ -58,9 +58,9 @@ class RawActivity(BaseModel):
     distance_meters: Optional[float] = None
     elevation_gain_meters: Optional[float] = None
 
-    # Heart rate
-    average_hr: Optional[int] = None
-    max_hr: Optional[int] = None
+    # Heart rate (Strava returns floats)
+    average_hr: Optional[float] = None
+    max_hr: Optional[float] = None
     has_hr_data: bool = False
 
     # User input
@@ -193,9 +193,9 @@ class NormalizedActivity(BaseModel):
     distance_meters: Optional[float] = None
     elevation_gain_m: Optional[float] = None
 
-    # Heart rate (optional)
-    average_hr: Optional[int] = None
-    max_hr: Optional[int] = None
+    # Heart rate (optional, Strava returns floats)
+    average_hr: Optional[float] = None
+    max_hr: Optional[float] = None
     has_hr_data: bool = False
 
     # User notes
